Refactor: Migrate KeyPairRepositoryImpl tests to use TestCoroutineDispatcher
- Replaced `runBlocking` with `runTest` and `StandardTestDispatcher` in `KeyPairRepositoryImplTest`.
- Added `kotlinx-coroutines-test` dependency.
- Removed unused `KeyPairRepositoryImpl.kt` file.

This change ensures that tests for `KeyPairRepositoryImpl` are executed using a more appropriate coroutine dispatcher for testing, improving test reliability and control over coroutine execution.
